requirements.txt
langchain 
langchain_ollama
selenium
beautifulsoup4
groq
python-dotenv